refactor: pull erc20abi to test files
