+++ test/automator-autoconverter.test.ts
@@ -26,9 +26,9 @@ import {
 } from "./utils";
 import { IVotingEscrow } from "../typechain/relay-private/lib/contracts/contracts/interfaces/IVotingEscrow";
 import { IERC20 } from "../typechain/openzeppelin-contracts/contracts/token/ERC20/IERC20";
-import { abi as erc20Abi } from "../web3-functions/relay-automation/abis/erc20.json";
 import { AutoConverterFactory } from "../typechain/relay-private/src/autoconverter";
 import { Registry } from "../typechain/relay-private/src";
+import { abi as erc20Abi } from "../test/erc20.json";
 
 import {
   KEEPER_REGISTRY_ADDRESS,
