+++ test/automator.test.ts
@@ -27,8 +27,8 @@ import {
 import { IVotingEscrow } from "../typechain/relay-private/lib/contracts/contracts/interfaces/IVotingEscrow";
 import { IERC20 } from "../typechain/openzeppelin-contracts/contracts/token/ERC20/IERC20";
 import { AutoCompounderFactory } from "../typechain/relay-private/src/autoCompounder";
-import { abi as erc20Abi } from "../web3-functions/relay-automation/abis/erc20.json";
 import { Registry } from "../typechain/relay-private/src";
+import { abi as erc20Abi } from "../test/erc20.json";
 
 import {
   KEEPER_REGISTRY_ADDRESS,
